% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{build_model_from_template}
\alias{build_model_from_template}
\title{builds a model object from a template which is given in a list of templates.}
\usage{
build_model_from_template(template_name, model_name = NULL,
  features = NULL, templates = default_templates, metric = "gini")
}
\arguments{
\item{template_name}{\code{character}: Specify template name. Must be among the names of the list given via argument \code{templates}}

\item{model_name}{\code{character}: Specify the name of the model being generated. Every model object can have a name. If name is not specified, a random name will be generated.}

\item{features}{\code{character or data.frame}: Specify feature names from which model will be trained. 
The models's config parameter \code{features.include} will be set accordingly to extract specified features from the training dataset.
features can also be a data frame of features (output of function evaluate_features)}

\item{templates}{\code{list}: A named list of templates. Argument \code{template_name} must be among the names of this list.}
}
\value{
A object inheriting from class \code{MODEL}
}
\description{
builds a model object from a template which is given in a list of templates.
}
