% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{correlation}
\alias{correlation}
\title{Use this function to get correlation of two vectors with various metrics.}
\usage{
correlation(x, y, metric = "pearson", threshold = NULL, ratio = NULL)
}
\value{
\code{numeric} the value of correlation between two given vectors by the specified metric
}
\description{
Use this function to get correlation of two vectors with various metrics.
}
\section{Fields}{

\describe{
\item{\code{x}}{\code{vector} first vector}

\item{\code{x}}{\code{vector} second vector (must have the same number of elements as vector \code{x})}

\item{\code{metric}}{\code{character} specifying metric. Valid values are
\code{pearson, loss, aurc, gini, 'precision', 'recall', 'f1', 'accuracy', 'sensitivity', 'specificity', 'fp', 'fn', 'tp', 'tn', 'lift'}}
}}

