% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{digits}
\alias{digits}
\title{5000 Handwritten digits}
\format{A list with 5000 images and their classification}
\usage{
data(digits)
}
\description{
A dataset containing 5000 20x20 pixel images of handwritten digits
The dataset comes from the Standford ML course 2014. It is a list of tow items:
1) a 5000 x 400 matrix containing the grey scale dn values for each image
in each row and 2) the classification of each digit as 0 - 9 as c(10,1-9).
}
\details{
\itemize{
  \item listitem1. a 5000 x 400 matrix
  \item listitem1. a 5000 x 1 vector of classifications
}
}
\keyword{datasets}

