% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{fit_models}
\alias{fit_models}
\title{Trains a given list of model objects and returns a trained list of models}
\usage{
fit_models(models, X, y, num_cores = 1, verbose = 1,
  remove_failed_models = T)
}
\description{
Trains a given list of model objects and returns a trained list of models
}
