% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logisPred}
\alias{decPlot}
\alias{logisPred}
\title{Prediction function for a fitted logistic classifier}
\usage{
logisPred(theta, X, thres = 0.5, type = "response")

decPlot(theta, dX, X, y, thres = 0.5)
}
\arguments{
\item{theta}{numeric vector of parameters}

\item{X}{design matrix: an array of numeric variables or features
including intercepts}

\item{thres}{decision boundry}

\item{type}{type of predictions on logit scale ("logit") or
"reponse" or as classification ("class")? Defaults to response}

\item{theta}{numeric vector of parameters}

\item{dX}{design matrix: an array of numeric variables or features}

\item{X}{original variables or features}

\item{y}{original classification
including intercepts}

\item{thres}{decision boundry}
}
\description{
Classifies from variables X

Classifies from variables X
}
\examples{
X<-array(runif(2000),dim=c(1000,2))
y<-sqrt(2*X[,2]^2+2.4*X[,1]^2)<.5
designX<-mapFeat(X)
designX<-cbind(rep(1,100),designX)
theta<-rep(0,ncol(designX))
lambda<-1
logisCost(theta,designX,y,lambda)
logisGrad(theta,designX,y,lambda)
par<-optim(theta,logisCost,logisGrad,X=designX,y=y,lambda=.01,method="BFGS")$par
decPlot(par,designX,X,y,thres=.5)
}
\author{
Marco D. Visser

Marco D. Visser
}

