% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mapFeat}
\alias{mapFeat}
\title{Feature mapping function to polynomial feautures}
\usage{
mapFeat(X, order = 2)
}
\arguments{
\item{X}{an array to be mapped. Must be numeric!}

\item{order}{the final polynomial order to map
(defaults to 2), may take a long time if order is large.
at this time also only accepts a value of two.}
}
\description{
mapFeat maps the input variable X into quadratic feautures.
for instance, if the input is a two dimensional array the function
returns a new array with more features, comprising of
  X1, X2, X1.^2, X2.^2, X1*X2, X1*X2.^2
}
\examples{
mapFeat(array(rnorm(20),dim=c(10,2)))
}
\author{
Marco D. Visser
}

