% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mClass}
\alias{digitAnim}
\alias{mClass}
\alias{mPred}
\title{m- Class classifier}
\usage{
mClass(trainingdata, y, lambda = 1)

mPred(Data, Theta)

digitAnim(Data, Theta, y, maxit = 10)
}
\arguments{
\item{trainingdata}{numeric matrix with training data}

\item{y}{classes corresponding to the rows of trainingdata}

\item{lambda}{regularization coefficient}

\item{Data}{numeric matrix with data to classify}

\item{Theta}{a previously trained classifier}

\item{Data}{numeric matrix with data to classify}

\item{Theta}{a previously trained classifier}

\item{maxit}{maximum number of random displayes}

\item{y}{True classes used in training}
}
\description{
Trains a m-class classifer from dataset trainingdata

uses a trained classifier to predict classes

uses a trained classifier to predict classes
and displayes random images with predictions on
screeen with a delay
}
\examples{
data(digits)
X<-digits[[1]]
y<-digits[[2]]
trained<-mClass(X,y)
data(digits)
X<-digits[[1]]
y<-digits[[2]]
trained<-mClass(X,y)
pred<-mPred(X,trained)
image1200<-t(as.array(X[1200,]))
pred1200<-mPred(image1200,trained)
## force correct rotation of matrix image
f <- function(m) t(m)[,nrow(m):1]
image(f(matrix(X[1200,]),ncol=20,nrow=20),main=pred1200)
data(digits)
X<-digits[[1]]
y<-digits[[2]]
trained<-mClass(X,y)
}
\author{
Marco D. Visser

Marco D. Visser

Marco D. Visser
}

