% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\docType{class}
\name{MODEL-class}
\alias{MODEL-class}
\alias{MODEL}
\title{Abstract Model Class}
\description{
Reference Class \code{MODEL} is the highest level of herarchy in a series of classes defined in the \code{rml} package.
All transformers, classifiers, regressors and survival models inherit from this class.
}
\section{Fields}{

\describe{
\item{\code{name}}{character containing name of the model}

\item{\code{type}}{character containing model type. Specifies transformer type}

\item{\code{package}}{character specifies name of the package model is using}

\item{\code{package_language}}{character specifies programming language of the package either R or Python.}

\item{\code{description}}{character descripbes what the model is.}

\item{\code{config}}{list containing all parameters and settings of the model. Values depend on the model class.}

\item{\code{fitted}}{logical specifies if the model is fitted (trained) or not}

\item{\code{transformers}}{list containing all transformer objects. Each transformer is an instance inheriting from MODEL class.}

\item{\code{gradient_transformers}}{list containing all gradient transformer objects. Each gradient transformer is an instance inheriting from MODEL class.}

\item{\code{yin_transformers}}{list of functions.}

\item{\code{yout_transformers}}{list of functions.}

\item{\code{objects}}{list containing additional data}

\item{\code{objects$features}}{data.frame containing list of features and their characteristics. This table is created after fitting the model.}

\item{\code{objects$fitting_time}}{POSIXct containing time when the model's training is complete. Created after the model is fitted.}
}}


