% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{model_make_unique_transformer_names}
\alias{model_make_unique_transformer_names}
\title{Changes duplicated transformer names of a given model to ensure all transformers have distinct names}
\usage{
model_make_unique_transformer_names(model)
}
\value{
None
}
\description{
Changes duplicated transformer names of a given model to ensure all transformers have distinct names
}
\section{Fields}{

\describe{
\item{\code{model}}{Any object inheriting from class \code{MODEL}}
}}

