% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{model_types}
\alias{model_types}
\title{Returns model type and types of its transformers recursively}
\usage{
model_types(model)
}
\value{
\code{character} vector containing model type and all its transformer types
}
\description{
Returns model type and types of its transformers recursively
}
\section{Fields}{

\describe{
\item{\code{model}}{Any object inheriting from class \code{MODEL}}
}}

