% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{moment}
\alias{moment}
\title{Returns desired moments of a given numeric vector}
\usage{
moment(v, m = c(1, 2), na.rm = T)
}
\value{
\code{numeric} output vector containing moment values
}
\description{
Returns desired moments of a given numeric vector
}
\section{Fields}{

\describe{
\item{\code{v}}{\code{numeric} input vector for which moments are to be computed}

\item{\code{m}}{\code{integer} desired moment orders}

\item{\code{na.rm}}{\code{logical} should missing values be removed?}
}}

