% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{na2median}
\alias{na2median}
\title{Imputes missing values with the median of non-existing values in each column}
\usage{
na2median(X)
}
\value{
\code{data.frame} with missing values replaced by imputed values
}
\description{
Imputes missing values with the median of non-existing values in each column
}
\section{Fields}{

\describe{
\item{\code{X}}{\code{data.frame} or \code{matrix}: Input table to be imputed}
}}

