% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{outlier}
\alias{outlier}
\title{This function gets a vector and returns a vector of logical which is TRUE if the corresponding value is an outlier}
\usage{
outlier(x, sd_threshold = 4, recursive = F)
}
\value{
logical vector specifying which element of the given vector \code{x} is outlier.
}
\description{
This function gets a vector and returns a vector of logical which is TRUE if the corresponding value is an outlier
}
\section{Fields}{

\describe{
\item{\code{x}}{\code{numeric} : Input vector}

\item{\code{sd_threshold}}{\code{numeric} : Standard Deviation threshold as a criteria for outlier determination.}

\item{\code{recursive}}{\code{logical} : TRUE if the process of outlier detection should continue with eliminated outliers}
}}

