% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvis.R
\name{plot_bindensity}
\alias{plot_bindensity}
\title{This function plots density of vector \code{x} for elements for which \code{y} is 0 in black and
another density plot for elements for which \code{y} is 1 in red.}
\usage{
plot_bindensity(x, y)
}
\description{
This function plots density of vector \code{x} for elements for which \code{y} is 0 in black and
another density plot for elements for which \code{y} is 1 in red.
}
