% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{predict_models}
\alias{predict_models}
\title{Calls \code{predict} method of each model in the given list and retunrs results in a data.frame}
\usage{
predict_models(models, X, num_cores = 1, verbose = 1)
}
\description{
Calls \code{predict} method of each model in the given list and retunrs results in a data.frame
}
