% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{ranker}
\alias{ranker}
\title{This function gets a data.frame and replaces each cell with its rank among all the cells in its column.}
\usage{
ranker(X)
}
\value{
\code{data.frame} containing ranked values. Suffix \code{_rank} is added to each column header.
}
\description{
This function gets a data.frame and replaces each cell with its rank among all the cells in its column.
}
\section{Fields}{

\describe{
\item{\code{X}}{\code{data.frame} or \code{matrix}: Input table to be ranked}
}}

