% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{scorer}
\alias{scorer}
\title{Computes various predictive performance scores between two binary columns of a table}
\usage{
scorer(df, prediction_col, actual_col)
}
\value{
\code{list} containing various scores. Returned scores are:
tp (True Positive), tn (True Negative), fp (False Positive), fn (False Negative),
tpr (True Positive Rate) or recall ro sensitivity or hit_rate,
fnr (False Negative Rate) or miss_rate,
tnr (True Negative Rate) or specificity or selectivity
fpr (False Positive Rate) or fall_out,
ppv (Positive Predictive Value) or precision
fdr (False Discovery Rate),
for (False Omission Rate),
pt (Prevalence Threshold),
ts (Threat Score) or csi (Critical Success Index)
accuracy, ba (Balanced Accuracy)
f1 (F1 Score), fmi (Fowlkes–Mallows Index),
informedness, markedness
mcc (Matthews Correlation Coefficient)
lift
}
\description{
Computes various predictive performance scores between two binary columns of a table
}
\section{Fields}{

\describe{
\item{\code{df}}{\code{data.frame, tibble, data.table} or \code{matrix}: Input table containing at least two columns with binary data}

\item{\code{prediction_col}}{\code{character} header of the column containing predicted binary classes}

\item{\code{actual_col}}{\code{character} header of the column containing actual binary classes}
}}

