% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{scorer_downsampled}
\alias{scorer_downsampled}
\title{Computes actual performance scores assuming the test set is down-sampled for class 0, by 'weight' times
The precision is expected to be lower than scorer, but recall remains the same.}
\usage{
scorer_downsampled(tbl, prediction_col, actual_col, weight = 2)
}
\description{
Computes actual performance scores assuming the test set is down-sampled for class 0, by 'weight' times
The precision is expected to be lower than scorer, but recall remains the same.
}
